refactor: moving the eventhub client
