+++ azurerm/internal/provider/services.go
@@ -21,6 +21,7 @@ import (
 	"github.com/terraform-providers/terraform-provider-azurerm/azurerm/internal/services/devtestlabs"
 	"github.com/terraform-providers/terraform-provider-azurerm/azurerm/internal/services/dns"
 	"github.com/terraform-providers/terraform-provider-azurerm/azurerm/internal/services/eventgrid"
+	"github.com/terraform-providers/terraform-provider-azurerm/azurerm/internal/services/eventhub"
 )
 
 func SupportedServices() []common.ServiceRegistration {
@@ -44,5 +45,6 @@ func SupportedServices() []common.ServiceRegistration {
 		devtestlabs.Registration{},
 		dns.Registration{},
 		eventgrid.Registration{},
+		eventhub.Registration{},
 	}
 }
